% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debrief.R
\name{check_pkg_available}
\alias{check_pkg_available}
\title{Check if required package is available}
\usage{
check_pkg_available(pkg, reason)
}
\arguments{
\item{pkg}{Package name}

\item{reason}{Description of why the package is needed}
}
\description{
Check if required package is available
}
\keyword{internal}
