% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debrief.R
\name{create_failure_records}
\alias{create_failure_records}
\title{Create standardized failure records}
\usage{
create_failure_records(failures, step, row_id_col)
}
\description{
Create standardized failure records
}
\keyword{internal}
