% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debrief.R
\name{process_db_failures}
\alias{process_db_failures}
\title{Process failures for database tables}
\usage{
process_db_failures(
  failure_query,
  step,
  chunk_size,
  output_to_parquet,
  parquet_path,
  output_to_db,
  con,
  failure_table_name,
  output_to_df,
  row_id_col
)
}
\description{
Process failures for database tables
}
\keyword{internal}
